requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.1.3
django-admin-honeypot==1.1.0
django-cors-headers==3.7.0
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
drf-yasg==1.20.0
Faker==4.14.2
flake8==3.9.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
validate-docbr==1.8.2
yapf==0.31.0
